Fix reference to global namespace class
Hi! I've found this bug that can be replicated by setting up the project (or using the 2.0.x branch of the Demo repo), activating the `wp-steroids` plug-in, navigating to the Wordpress permalink config page and clicking on "save".

It complains that cannot find the  `WPS_Object_Cache` class that is loaded within the `wp-steroids` plug-in.

By prefixing the class with a backslash it works, as it now correctly picks it up from the global namespace.